Better slide menu
I uprated the slide menu, but only one button for a test now and I also added new function which generates the trend for humidity, but the code was just to make it work. In the futre I think i will make one function for all trend buttons and there will be a case inside it which decides which btn was clicked.